% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bodyfat.R
\name{bodyfat}
\alias{bodyfat}
\title{Body Fat Percentage}
\usage{
bodyfat(weight, wrist, waist, hip, forearm)
}
\value{
This function will return the lean body mass and body fat percentage of someone who
is a woman.
}
\description{
This function, bodyfat, takes all the measurements needed to determine the body fat
percentage of a woman and calculate both the lean body mass and body fat percentage of
that said person.
}
\details{
important
This function will only calculate the lean body mass and body fat percentage with
measurements for a women.

weight <- The weight in pounds of that person.
wrist <- The circumference in inches of the person's wrist, at the widest spot,
 divided by the correct constant for women.
waist <- The circumference in inches of the person's waist, at the naval,
 multiplied by the correct constant for women.
hip <- The circumference in inches of a person's hips, at their fullest point,
 multiplied by the correct constant for women.
forearm <- The circumference in inches of a person's forearms, at the widest point,
 multiplied by the correct constant for women.
LBM <- The caculation to determine the lean body mass of that person with their
 weight and wrist, waist, hip, and forearm circumferences being considered.
fat_percent <- The caculation of the body fat percentage of that person with their
 previously calculated lean body mass and weight being considered.
}
