% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendarHeatmap.R
\name{calendarHeatmap}
\alias{calendarHeatmap}
\title{Calendar Heatmap}
\usage{
calendarHeatmap(
  dates,
  values,
  title = "",
  subtitle = "",
  legendtitle = "",
  ncol = 1
)
}
\arguments{
\item{dates}{A vector containing the dates in \code{Date} format.}

\item{values}{A vector containing the corresponding values as numeric.}

\item{title}{Main plot title (optional).}

\item{subtitle}{Main plot subtitle (optional).}

\item{legendtitle}{Legend title (optional).}

\item{ncol}{number of columns in the plot (optional); default: ncol = 1.}
}
\value{
ggplot object
}
\description{
Creates a colour coded calendar visualising time series data
}
\details{
Taken from \url{https://quantixed.org/2022/12/26/twitcher-ii-tweet-frequency-and-top-tweets/#heatmap} and
\url{https://www.r-bloggers.com/2023/01/running-around-2022-running-dataviz-in-r/}
}
