% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calories_burned_boxpt.R
\name{calories_burned_boxpt}
\alias{calories_burned_boxpt}
\title{Calories Burned in Each Workout Day Boxplot}
\usage{
calories_burned_boxpt(data, x, y)
}
\value{
calories_burned_boxpt will return an aesthetic colored boxplot of the distribution
of the calories burned in each workout day.
}
\description{
The function, calories_burned_boxpt, takes an x-variable that is categorical and
numeric y-variable of a dataset to create a boxplot in ggplot.
}
\details{
data <- The desired data set being used
x <- The workout day that is being examined.
y <- The amount of calories burned

The ggplot() command maps the x-variable and y-varialbe from the data set desired
to use for the plot.
The geom_boxplot command maps the color of each level in the x-variable to be
specified through scale_color_manual and for the inside of each box-whisker to be
be colored to be gray70 throught scale_fill_manual command.

A x-axis, y-axis, and title for the graph are added generically where whatever
x-variable and y-variable is desired will be printed out on the graph.
}
