% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createUebersicht.R
\name{createUebersicht}
\alias{createUebersicht}
\title{Erzeuge eine Ueberischt der Fitnessdaten}
\usage{
createUebersicht(data, sportart = "Indoor Cycling")
}
\arguments{
\item{data}{data.frame mit Messwerten (im richtigen Format)}

\item{sportart}{character-String; aktuell implementiert: "Indoor Cycling" (default),
"Cycling", "Walking", "Running"}
}
\value{
uebersicht data.frame mit der Übersicht
}
\description{
Erzeuge eine Ueberischt der Fitnessdaten
}
\author{
Ulf Schepsmeier
}
