% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectTypeOfSport.R
\name{detectTypeOfSport}
\alias{detectTypeOfSport}
\title{.fit Dateien erkennen und einer Sportart zuordnen}
\usage{
detectTypeOfSport(path, file)
}
\arguments{
\item{path}{character-String mit dem Pfad zu dem fit-File}

\item{file}{character-String mit dem Namen der fit-Datei}
}
\value{
sportart
}
\description{
Nach einer Umstellung von Wahoo ist es schwieriger die .fit Dateien in Dropbox an die richtige Stelle zu tun.
Aktuell passiert das manuell. Um das zu automatisieren muss ich automatisch erkennen, ob es sich um
indoor Cycling, Cycling, walking oder running Daten handelt, um die .fit Datei in den richtigen Ordner zu kopieren.
}
\author{
Ulf Schepsmeier
}
