% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit2csv.R
\name{fit2csv}
\alias{fit2csv}
\title{.fit Dateien umwandeln/auslesen}
\usage{
fit2csv(path, file, outpath = NULL)
}
\arguments{
\item{path}{character-String mit dem Pfad zu dem fit-File}

\item{file}{character-String mit dem Namen der fit-Datei}

\item{outpath}{character-String mit dem Pfad für den Output; default: outpath = NULL, i.e. the fit-path from 'path'
is replaced with 'fixed'}
}
\value{
boolean ob die Umwandlung funktioniert hat oder nicht
}
\description{
.fit Dateien umwandeln/auslesen
}
\note{
Das Programm GPSBabel muss installiert sein
}
\author{
Ulf Schepsmeier
}
