% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit2csv.R
\name{fit2csvOrdner}
\alias{fit2csvOrdner}
\title{Transformiere einen ganzen Ordner vom .fit-Format in CSV}
\usage{
fit2csvOrdner(ordner, path)
}
\arguments{
\item{ordner}{character-String der Ordners}

\item{path}{charater-String mit absolutem Pfad bis zum Ordner
(Leider ist der absolute Pfad notwendig für die EXE im Hintergrund)}
}
\description{
Transformiere einen ganzen Ordner vom .fit-Format in CSV
}
\author{
Ulf Schepsmeier
}
