% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitnessPlot.R
\name{FormPlot}
\alias{FormPlot}
\title{Form plot}
\usage{
FormPlot(df, toStr)
}
\arguments{
\item{df}{data.frame mit den Aufbereiteten Daten aus der Funktion calculateTL()}

\item{toStr}{character; Enddatum in der Form "\%Y-\%m-\%d"}
}
\value{
ggplot-Objekt
}
\description{
Form plot
}
\note{
Date Daten werden von einem Jahr berechnet, aber nur die letzten 180 Tage dargestellt
}
\author{
Ulf Schepsmeier
}
