% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getElevation.R
\name{getElevation}
\alias{getElevation}
\title{getElevation}
\usage{
getElevation(df)
}
\arguments{
\item{df}{data.frame mit latitude und longitude}
}
\value{
data.frame mit latitude, longitude und elevation
}
\description{
Nach einer Umstellung von Wahoo sind in der .fit-Datei keine Angaben mehr zur Elevation (Höhe) von Koordinaten.
Deswegen hole ich mir diese Informationen über eine open-source API.
}
\note{
Ich benutze die open-source API von \url{https://open-elevation.com/}
Details zu der Benutzung kann der README Seite von
\url{https://github.com/Jorl17/open-elevation/blob/master/docs/api.md} entnommen werden.
}
\author{
Ulf Schepsmeier
}
