% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFitnessData.R
\name{getFitnessData}
\alias{getFitnessData}
\title{Einlesen der Fitness-Daten aus den CSV Dateien}
\usage{
getFitnessData(ordner, path, skip = NULL)
}
\arguments{
\item{ordner}{character-String mit dem Ordner-Namen
"walking", "indoorCycling", "running" oder "cycling"}

\item{path}{character-string mit dem Pfad-Namen bis zu dem Ordner}

\item{skip}{vector mit Indices mit zu überspringenden Dateien in dem Ordner;
default: skip = NULL}
}
\value{
Data eingelesene Daten
}
\description{
Einlesen der Fitness-Daten aus den CSV Dateien
}
\author{
Ulf Schepsmeier
}
