% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indoorCycling.R
\name{getIndoorCyclingData}
\alias{getIndoorCyclingData}
\title{Auslesen und Aufbereiten der Daten fuer Indoor Cycling und Cycling}
\usage{
getIndoorCyclingData(path, file)
}
\arguments{
\item{path}{character-String mit dem Pfad zu dem CSV-File}

\item{file}{character-String mit dem Namen des CSV-Files}
}
\value{
rawData data.frame mit den aufbereiteten Daten
}
\description{
Auslesen und Aufbereiten der Daten fuer Indoor Cycling und Cycling
}
