% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indoorCycling.R
\name{getSummaryIndoorCycling}
\alias{getSummaryIndoorCycling}
\title{Erzeuge eine Zusammenfassung der Daten fuer Indoor Cycling oder Cycling}
\usage{
getSummaryIndoorCycling(indoorCyclingData)
}
\arguments{
\item{indoorCyclingData}{data.frame mit den Fitness-Daten}
}
\value{
summaryIndoorCycling data.frame
}
\description{
Erzeuge eine Zusammenfassung der Daten fuer Indoor Cycling oder Cycling
}
