% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walking.R
\name{getSummaryWalking}
\alias{getSummaryWalking}
\title{Erzeuge eine Zusammenfassung der Daten fuer Walking oder Running Daten}
\usage{
getSummaryWalking(walkingData)
}
\arguments{
\item{walkingData}{data.frame mit den selektierten Daten}
}
\value{
summaryWalking data.frame mit den aggregierten Daten
}
\description{
Erzeuge eine Zusammenfassung der Daten fuer Walking oder Running Daten
}
\author{
Ulf Schepsmeier
}
