% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walking.R
\name{getWalkingData}
\alias{getWalkingData}
\title{Auslesen und Aufbereitung der Daten fuer Walking und Running}
\usage{
getWalkingData(path, file)
}
\arguments{
\item{path}{character-String mit dem Pfad zu der Datei}

\item{file}{character-String mit dem Namen der csv-Datei}
}
\value{
rawData data.frame mit den eingelesenen und aufbereiteten Daten
}
\description{
Auslesen und Aufbereitung der Daten fuer Walking und Running
}
\author{
Ulf Schepsmeier
}
