% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtJahresvergleich.R
\name{gtJahresvergleichFct}
\alias{gtJahresvergleichFct}
\title{gt-Tabelle für Jahresvergleich}
\usage{
gtJahresvergleichFct(data, type = "Anzahl", short = "#")
}
\arguments{
\item{data}{data.frame mit aggregierten Daten aus den verschiedenen Jahren, s. Note}

\item{type}{string; Für welche Variable soll die Auswertung gemacht werden? "Anzahl" oder "km";
default: type = "Anzahl"}

\item{short}{string; Spaltenkürzel in der finalen Tabelle; default: short = "#"}
}
\value{
gt-Objekt
}
\description{
Im Jahresvergleich möchte ich die Anzahl bzw. die km von mehreren Jahren vergleichen.
Es werden Differenzen zum Vorjahr berechnet und negative rot hervorgehoben, sowie Gesamt-Anzahl bzw. Gesamt-km.
}
\note{
Der data.frame muss einen bestimmten Aufbau und damit bestimmte Spalten haben:
\itemize{
\item Jahr \if{html}{\out{<chr>}} mit Zeilen für die Jahre + "Gesamt" (pro Sportart)
\item Sportart \if{html}{\out{<chr>}} bei mir "Cycling" und "Indoor Cycling"
\item Anzahl \if{html}{\out{<int>}}
\item km \if{html}{\out{<dbl>}}
\item DiffAnzahl \if{html}{\out{<dbl>}} Änderung zum Vorjahr; Bei "Gesamt" 0
\item DiffKm \if{html}{\out{<dbl>}} Änderung zum Vorjahr; Bei "Gesamt" 0
\item JahrInt \if{html}{\out{<dbl>}} Das "Jahr" als numerischer Wert; "Gesamt" = 0
}
}
\author{
Ulf Schepsmeier
}
