% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadAndFitness.R
\name{makeDateDF}
\alias{makeDateDF}
\title{Make a data frame that has every day in our time window represented}
\usage{
makeDateDF(fromStr, toStr)
}
\arguments{
\item{fromStr}{character; Startdatum in der Form "\%Y-\%m-\%d"}

\item{toStr}{character; Enddatum in der Form "\%Y-\%m-\%d"}
}
\value{
df
}
\description{
Make a data frame that has every day in our time window represented
}
\note{
Based on [www.r-bloggers.com](https://www.r-bloggers.com/2022/11/form-and-file-estimating-running-form-in-r/)
}
\author{
Ulf Schepsmeier
}
