% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indoorCycling.R
\name{plotIndoorCycling}
\alias{plotIndoorCycling}
\title{Plot fuer Indoor Cycling oder Cycling Daten}
\usage{
plotIndoorCycling(rawData, type = "IndoorCycling")
}
\arguments{
\item{rawData}{data.frame mit den Fitness-Daten}

\item{type}{character-String; "IndoorCycling" (default) oder "Cycling}
}
\value{
ggplot2-Objekt
}
\description{
Plot fuer Indoor Cycling oder Cycling Daten
}
\note{
Der Plot für die Herzrate hat jetzt als neues Feature noch die Herzratzonen farblich hinterlegt.
Die Zonen basieren auf meinen Daten (männlich, 35 Jahre, 80km, 1.86m)
}
