% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walking.R
\name{plotWalking}
\alias{plotWalking}
\title{Plot fuer Walking oder Running Daten}
\usage{
plotWalking(rawData, type = "Walking")
}
\arguments{
\item{rawData}{data.frame mit den selektierten Daten}

\item{type}{charater-String; "Walking" (default) oder "Running"}
}
\value{
p ggplot2-Objekt
}
\description{
Plot fuer Walking oder Running Daten
}
\author{
Ulf Schepsmeier
}
