% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadAndFitness.R
\name{sumDays}
\alias{sumDays}
\title{Sum the load for each day}
\usage{
sumDays(df, daydf)
}
\arguments{
\item{df}{data.frame with the load (calculated with calculateLoad())}

\item{daydf}{data.frame from makeDateDF()}
}
\value{
newdf merged data.frame
}
\description{
Sum the load for each day
}
\author{
Ulf Schepsmeier
}
