% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadAndFitness.R
\name{calculateLoad}
\alias{calculateLoad}
\title{Berechne den Load}
\usage{
calculateLoad(df)
}
\arguments{
\item{df}{data.frame der Uebersichtstabelle (aggregierte Daten pro Training)}
}
\value{
df
}
\description{
Funktion um die Last (Load) zu berechnen
}
\details{
The idea behind training load is to have a metric how a training has an effect to your body.
Second, training load is (in theory) independent of the sport type or the intensity.
Thereby, we are able to compare a short but intensive training with a long but
low intense training.

Since I do not have a power meter I have to estimate the load and the training stress score (TSS)
based on heart rate. Then it is called hrTSS.
see \url{https://help.trainingpeaks.com/hc/en-us/articles/204071944-Training-Stress-Scores-TSS-Explained}
hrTSS has some shortcomings \cr
Quote from \url{https://de-eu.wahoofitness.com/blog/ate-what-is-tss/} \cr
"Heart rate TSS or hrTSS is a less accurate model for determining your TSS of a workout.
While hrTSS is accurate when performing steady-state efforts, or endurance rides, it is far less accurate when
intervals come into play. \cr
hrTSS is based on time spent in each heart rate zone based on your Threshold Heart Rate. \cr
When there are large or frequent fluctuations in your heart rate data, such as data post-interval session,
it is hard for an accurate TSS to be quantified from this as it takes time for your HR to stabilize pre,
during, and post effort."
}
\note{
add a column that contains the load of each activity
one way to calculate load is to multiply time in hours by avg HR and add 2.5 times avg HR
this relates to load by y = ax + b of a = 0.418, b = -150 \cr
These numbers are taken from \url{https://www.r-bloggers.com/2022/11/form-and-file-estimating-running-form-in-r/}
and have to be adjusted for me (TODO)
}
\references{
\url{https://www.r-bloggers.com/2022/11/form-and-file-estimating-running-form-in-r/}
}
\author{
Ulf Schepsmeier
}
