% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadAndFitness.R
\name{calculateTL}
\alias{calculateTL}
\title{calculate training loads}
\usage{
calculateTL(df)
}
\arguments{
\item{df}{data.frame with the loads (after sumDays())}
}
\value{
df data.frame with calculated values for Fitness (CTL), Fatigue (ATL) and Form (TSS)
}
\description{
calculate training loads
}
\note{
Based on [www.r-bloggers.com](https://www.r-bloggers.com/2022/11/form-and-file-estimating-running-form-in-r/)
}
\author{
Ulf Schepsmeier
}
