% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createUebersicht.R
\name{createUebersicht}
\alias{createUebersicht}
\title{Erstellung einer Kennzahlenuebersicht}
\usage{
createUebersicht(data, sportart = "Indoor Cycling")
}
\arguments{
\item{data}{data.frame mit Messwerten (im richtigen Format)}

\item{sportart}{character-String; aktuell implementiert: "Indoor Cycling" (default),
"Cycling", "Walking", "Running"}
}
\value{
uebersicht data.frame mit der Uebersicht
}
\description{
Erzeuge eine Ueberischt der Fitnessdaten (Summary) mit den wichtigsten (aggregierten) Kennzahlen
}
\note{
Die berechneten Kennzahlen sind:
\itemize{
\item Trainingszeit
\item Distanz
\item Durchschnittsgeschwindigkeit
\item Maximale Geschwindigkeit
\item Durchschnittsherzrate
\item Maximale Herzrate
\item Standardabweichung bei der Herzrate (für interne Nutzung)
\item Standardabweichung bei der Geschwindigkeit(für interne Nutzung)
\item Durchschnittstrittanzahl (bei Rad, wenn vorhanden)
\item Maximale Trittanzahl (bei Rad, wenn vorhanden)
\item Standardabweichung bei der Trittanzahl (bei Rad, wenn vorhanden)
}
}
\author{
Ulf Schepsmeier
}
