% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitnessPlot.R
\name{FitnessAndFatigurePlot}
\alias{FitnessAndFatigurePlot}
\title{Fitness and Fatigure plot}
\usage{
FitnessAndFatigurePlot(df, toStr)
}
\arguments{
\item{df}{data.frame mit den Aufbereiteten Daten aus der Funktion calculateTL()}

\item{toStr}{character; Enddatum in der Form "\%Y-\%m-\%d"}
}
\value{
ggplot-Objekt
}
\description{
Fitness and Fatigure plot
}
\note{
Date Daten werden von einem Jahr berechnet, aber nucr die letzten 180 Tage dargestellt
}
\author{
Ulf Schepsmeier
}
