% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEntwicklung.R
\name{plotEntwicklung}
\alias{plotEntwicklung}
\title{Plot um die Entwicklung ueber die Zeit darzustellen}
\usage{
plotEntwicklung(data)
}
\arguments{
\item{data}{data.frame mit Daten der Sportart, die dargestellt werden soll}
}
\value{
ggplotly-Objekt
}
\description{
Plot um die Entwicklung ueber die Zeit darzustellen
}
\author{
Ulf Schepsmeier
}
