% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadAndFitness.R
\name{sumDays}
\alias{sumDays}
\title{Aufsummieren der Loads}
\usage{
sumDays(df, daydf)
}
\arguments{
\item{df}{data.frame with the load (calculated with calculateLoad())}

\item{daydf}{data.frame from makeDateDF()}
}
\value{
newdf merged data.frame
}
\description{
Sum the load for each day
}
\references{
\url{https://www.r-bloggers.com/2022/11/form-and-file-estimating-running-form-in-r/}
}
\author{
Ulf Schepsmeier
}
